segment pane code into logical files
